+++ api/src/main/java/org/allaymc/api/form/element/Slider.java
@@ -24,6 +24,8 @@ public final class Slider implements CustomFormElement {
     @Getter
     @SerializedName("default")
     private float defaultValue;
+    @Getter
+    private final String tooltip;
 
     /**
      * Create a new slider.
@@ -34,12 +36,13 @@ public final class Slider implements CustomFormElement {
      * @param step         the step of the slider
      * @param defaultValue the default value of the slider
      */
-    public Slider(String text, float min, float max, int step, float defaultValue) {
+    public Slider(String text, float min, float max, int step, float defaultValue, String tooltip) {
         this.text = text;
         this.min = Math.max(min, 0f);
         this.max = Math.max(max, this.min);
         this.step = step;
         this.defaultValue = defaultValue;
+        this.tooltip = tooltip;
     }
 
     @Override